requirement.txt
flask
flask-cors
langchain
langchain-community
langchain-core
langchain-google-genai
faiss-cpu
youtube-transcript-api
googletrans==4.0.0-rc1
python-dotenv
